systemd::cache = false result was not vague.
Starting with

```puppet
class{ 'systemd':
  manage_reolved => true,
  cache          => false,
}
```

note that the cache parameter is currently `false` by default and an explicit
`Cache=no` would not be set. Instead nothing would be done.

This patch maintains backwards compatibility in that the default
still results in no modification to `/etc/systemd/resolved.conf`.

However an explicit setting of `cache => false` with this patch will now add
`Cache=no` to `resolved.conf`.